Add per-change flags and wallpaper windows to transitions
This CL adds flags per-change in a transition so that
extra meta-information (like is-wallpaper, shows-wallpaper,
translucent, starting-window-transfer) can be passed to
shell.

This then treats wallpaper like other activities WRT to
transitions (collect/send to shell) except for a couple
differences:
1. When any window is collected, if it shows wallpaper, then
   the wallpaper will be automatically collected too.
2. During TransitionInfo construction, wallpaper containers
   are excluded from promotion and root-leash calculation.

This also adds starting-window transfers to the transition
and adds support in the example player.

Bug: 169035082
Test: launch any non-wallpaper app (eg. messages), then press
      home-button to bring home to front. The wallpaper should
      be included with the home fading-in.
Change-Id: I7ceea614e16c2686f533937db10436f66764255e

Former-commit-id: dba8bbfd5e42094926c7f56c7290d5e7fa87e9f8