Add utils for converting Maps and ParcelUuid to/from PersistableBundle
This commit expands the PersistableBundleUtils by adding maps.
LinkedHashMap is used in an attempt to preserve ordering where stability
is important.

Similarly, this commit adds the ability to persist ParcelUuid(s) via
conversion to Strings.

This commit also adds a helper method to safely read and write to and
from disk

Bug: 163611304
Test: New tests added, passing
Change-Id: Ife24e94006445007be68ab0e03f27b2fd5643aa2

Former-commit-id: 296c895e10cf835e3158e0327e4f202559323f3e