Fix SystemUI count down doesn't work
StatusBarIconView uses Icon#loadDrawableAsUser to load drawable defined
by the customized Drawable class. context.createContextAsUser with flag
Context.CONTEXT_INCLUDE_CODE can fix the ClassNotFound found problem.
And, it needs to prevent any security issue by checking if the context
is the same app with the process.

If the parameter userId is the same with context.getUserId(), it
shouldn't create the extra Context instance.

Fixes: 171712189
Test: atest \
    CtsGraphicsTestCases:android.graphics.drawable.cts.IconTest \
    FrameworksCoreTests:android.graphics.drawable.IconTest
Change-Id: I267fc6eed90a6a6bba39b9976c1d843734e49dec

Former-commit-id: 920dcad2d34bdeb663779402bfe361cf8c160842