Remove @SystemApi from Compatibility
Bug: 152322291
Bug: 172934376
Test: make

The UI Rendering mainline module is on hold for now, so this does not
need to be @SystemApi. The system can still access it through @hide.

This leaves in place the new Compatibility class (added in
Ie7172fb93364a1e04ab844b8fa64887bf9d8b005), which seems just as good as
the old approach, and we can switch it back to @SystemApi if/when we
restart the module.

Change-Id: I3f6c32603c4eace55ece04936bff6f9336cf6627

Former-commit-id: 2af10003c42cc46525948ee11ed1a2a55b8a355d