Add assertions to CloseAppBackButtonTest
Adding all the assertions for the CloseAppBackButtonTest. This includes WM assertions (the common ones for all test + one that the launcher replaces the app window and another that the wallpaper window becomes visible) and layers assertions (the common ones + one for the wallpaper layer replacing the app layer).
Also adding new assertions in CommonAssertions files (wallpaperLayerReplacesAppLayer, wallpaperWindowBecomesVisible and launcherReplacesAppWindowAsTopWindow).

Test: atest FlickerTests:CloseAppBackButtonTest
Bug: b/171049647
Change-Id: I992edfe372b677d723dda09a0af8fdf9f6213c73

Former-commit-id: 042fac5bdf3cf24ad1b425b7704358b5a81b2b6a