StorageNotification.java: Apply FLAG_IMMUTABLE to various PIs
An explicit mutability flag is soon going to be required for any created
PendingIntent. go/immutable-pendingintents has context on this new
requirement. In this commit it seems reasonable to apply FLAG_IMMUTABLE
to these PIs for the following cases:

-- Simple notification content clicks are immutable; there's no need
to communicate customized data back to the creator
-- Simple notification action clicks are immutable; there's no need
to communicate customized data back to the creator

Bug: 173158147
Test: TH
Change-Id: I0525c067a3f18ae72edb52767b7d6807f2318444

Former-commit-id: 7b20a8f6c7ffe261ce980fecda717a86b37f62bb