Refaster templates for new TypedXml classes.
Since we have over 100 unique schemas across the OS, it would be
incredibly tedious to try migrating all that logic to use the more
efficient TypedXmlSerializer and TypedXmlPullParser interfaces.

To aid this migration process, this change adds Refaster templates
that offer to refactor matching existing code in a no-op fashion.

This change also upgrades Error Prone to the latest release.

Bug: 171832118
Test: manual
Change-Id: Ic93a46d646edee98af7d0e2f7891d931fca4825f

Former-commit-id: f14ed648a9e98d6a14bff12fd30faef77d7943db