Start home on empty display when user unlock the device
If the home app on primary display is encryption aware, the home app
will be launched during direct boot instead of FallbackHome. This may
cause the non-encryption aware home app not being launched on secondary
display after user unlock the device.

This CL try to start home on empty display again when user unlock the
device.

Bug: 171931543
Test: atest ActivityTaskSupervisorTests#testStartHomeAfterUserUnlocked
Change-Id: I5e4a2a7125bf49dc01ce5804de7596d29ded6d5c

Former-commit-id: d7aa077f2f224bc7a8b5999ce1b17c361653cb80