Add messageId to both of the MMS methods
There are two MMS-related methods in SmsManager, one to send a
multimedia message and one to download a multimedia message. In
R, a messageId parameter was added to the various SMS sending
methods. The messageId allows for eventual increased reliablility.
The messageId parameter was added to MmsManager for R, but
MmsManager was ultimately hidden in R. This change applies the
same treatment to the two MMS methods in S, that was applied to
the SMS methods in R.

Test: manual testing and atest MmsTest
Bug: 163413552

Merged-in: Ic8e59b652967ab1296a92a1d5727dc226cb190fb
Change-Id: Ic8e59b652967ab1296a92a1d5727dc226cb190fb

Former-commit-id: e8198d1dc009058df382721279e54afd7aa574f0