Stop copying VectorDrawable pixels when drawing.
Drawing a mutable SkBitmap to an SkCanvas will result in the pixels
being copied and a new genID assigned. This results in both a cpu
and gpu copy every time a VectorDrawable is redrawn.

The solution is to create an SkImage using the flag to instruct it
not to copy the pixels which is what Bitmap::makeImage() does.

Bug: 173732636
Test: captured SKP and verified the genID is consistent between frames
Change-Id: Ie13385e89de51c6b9ee2f2ba31eccbfdf3adac48

Former-commit-id: 2960d1693cb3842282663aa395708de788f92ffa