Add ability to rotate screen in AoD.
This patch prevents DisplayRotation from disabling the
OrientationListener when the screen turns off, if and only if the
underlying sensor is a wakeup device_orientation sensor (i.e. we don't
have to keep the AP on to keep listening). This then allows the AoD on
the device to rotate.

Since many devices doesn't have a wake orienation sensor, a config flag is
provided to force this behaviour on (although note that the AoD might
not rotate if the AP shuts down). This can be enabled by setting
config_forceOrientationListenerEnabledWhileDreaming to true.

Test: Not yet, finding tests...
Change-Id: Ie1ee8e297ebe9a8a73793d07f173bada8e0699f1

Former-commit-id: c7200e18ef28f64e76680c8a3073b0450c5df6b2