DO NOT MERGE Use round highlight for navigation buttons
Make the focus highlight and touch ripple for buttons in the navigation
bar be round rather than square.

Also fix a bug in the dimming logic. The alpha was applied to the icon
initially, but when the state changed, the alpha was applied to the
entire button.

Launcher in master is quite different so it this change is limited to
rvc-qpr-dev for now.

Test: manual
Bug: 174614908
Bug: 168317872
Change-Id: I9b80946887ad85a73ddd281f7afcc9fcc8280a64

Former-commit-id: 9242e6ee6a029a01635eb482b719dcd4d99fa157