Update last surface position before unrotating
Because the surface position will be set to the unrotated position
according to the last surface position, that also means the current
position is applied. Otherwise the reference position for seamless
rotator may be zero.

This also fixes testSeamlesslyRotateWindow didn't verify the exact
rotated position because WindowState#updateSurfacePosition was
changed to check isGoneForLayout.

Fixes: 174666072
Test: WindowStateTests#testSeamlesslyRotateWindow
Change-Id: I99105fcda9b2252a3a9fb50c10b9cba1a4f732f1

Former-commit-id: 836c889e257e9f6cbfaa2327bfc16730d1fbac81