Update rotary highlight on bordered buttons
This adds a rotary highlight around bordered
buttons. Previously the rotary highlight
was only visible as the ripple expanded to fill
the button, which was extremely subtle and
hard to see.

Due to issues with layering and corner radiuses,
this cl doesn't include the rotary fill color.
It only has the stroke around the outside
of the button. File b/175161842 to address
that.

Bug: 174490425
Test: Manually
Change-Id: I2abb54b9be04d33c642ce0509a88696afc028c9b

Former-commit-id: 8f65d7d249cb20d93c1e33b42cf306106ece451a