Fix Pip tests after Pip Menu migration to a Window
Instrumentation tests on TVs used to use UiObject2.click() method to
interact with UI elements. This was inject input events that would
prompt input framework to enable the touch mode on the device. This was
wrong as TV devices are not expecting to work in the touch mode.
Migration of Pip menu from an Activity to a Window exposed this issue,
which caused multiple TV Pip tests to fail.
With this CL we change the way we interact with UI: we only are using
DPad navigation, which also means that we have to "track" the focused
element.

Bug: 174818743
Test: atest WMShellFlickerTests:TvPipBasicTest
Test: atest WMShellFlickerTests:TvPipMenuTests
Test: atest WMShellFlickerTests:TvPipNotificationTests
Change-Id: I335f8fd8d9c840f751d207c5e60b6af582118fee

Former-commit-id: ecad63042b63a352f35b8660f0410502cdbbfcd4