[Attempt #2] Block untrusted touches in enforcing mode
ag/13253804 fixes test we broke in attempt #1 (ag/13228932).

-- Original CL description (update "Test:" line):

After getting enough feedback from dogfooders, the issues left open that
we know of (that will still be addressed ofc) either don't hurt dogfood
experience (eg. b/172627742 touch is blocked only during AGSA's -1
screen <> launcher side animation) or are 3p apps that will have to
change their code at some point. So, turning the feature on in its final
enforcing mode.

Also removing toast since at this point this is more noise than signal.
We're seeing many duplicate bugs (many of which are WAI) that the
dogfooder would likely not notice if it wasn't for the toast. If there
is a real issue the user will still notice by the unresponsiveness and
given a bugreport we will be able to check via logcat.

However, I have not removed the code completely and left that as part
of b/169067926 since we may decide to turn it back on before or on
DP/Beta.

Bug: 158002302
Test: atest WindowUntrustedTouchTest WindowInputTests inputflinger_tests
      inputflinger_benchmarks libinput_tests libgui_test
Test: Verify untrusted touches are blocked via logcat but no toast.
Change-Id: I709711e60f5f4f563e018ee079ef89418e161d14

Former-commit-id: 51a3a37b1bcde73e34800fe8aec0ae69a554856c