Add DomainVerificationCollector
Gathers the domains declared in a package's manifest under <activity>
entries for use with domain verification auto verification and user
selection.

The behavior is split between v1 and v2, for apps that target pre-S,
as the set of domains that are verified must be maintained until the
app author is able to update to the stricter S schema.

The S schema is defined as part of the JavaDoc on the @ChangeId, but
will also be included in the developer docs once these changes are
merged.

Exempt-From-Owner-Approval: Already approved by owners on main branch

Bug: 163565712

Test: atest DomainVerificationCollectorTest

Change-Id: I58c9d2b3001f48b9904aa4617d546a8cee0d926e

Former-commit-id: 9056557d347b33d04ed9f118fa6c47e01d076254