Use new always open during PMS Activity resolution
Implements the behavior of either delegating to an approved app for
a web domain, or the browsers, removing the "always ask" and "never"
options for applications.

This also considers cross profile intents, and will allow those to be
shown if the other profile has an approved handler.

This also migrates instant app state to the new behavior. Instant apps
will now be granted completely immutable autoVerify approval, to
preserve their automatic opening behavior. The user would have to
disable the app in settings similar to an auto approved system app.

This is flagged through USE_DOMAIN_VERIFICATION_V2. For now, both v1
and v2 run entirely parallel, until the v1 APIs are removed in a
follow up.

Exempt-From-Owner-Approval: Already approved by owners on main branch

Bug: 163564991

Test: TODO

Change-Id: I1a90f6b2b9bf2f108c5dc1b944985bc150c405a7

Former-commit-id: 0cab12964e3662ffef148243769b0520d090425a