Update notif onClick logic
NMS.onClick must be called before waiting for the shade to collapse so
that the correct cancellation reason is sent to NSLs (ie: REASON_CLICK = 1,
not REASON_APP_CANCEL = 8 or REASON_GROUP_SUMMARY_CANCELED = 12).

However, we also need to retrieve the group summary to dismiss before
NMS.onClick is sent to avoid a different race condition: b/171467963.

Test: atest SystemUITests
Fixes: 175518877
Change-Id: I7f2c3b0682fea3b56d940c62423cd8956192918d

Former-commit-id: 134206b120d2aad3b566d64b4ec5903fd1d19ad7