Combine v1 and v2 DomainVerficationProxies
In case both v1 and v2 are on the device, have a delegate that supports
sending requests to both receivers. The verification agent will then
decide which API to use.

This allows deferred migration to the v2 APIs. Note that the old user
state API is still non-functional, even for a v1 verification agent.

Exempt-From-Owner-Approval: Already approved by owners on main branch

Bug: 170321181

Test: atest DomainVerificationProxyTest

Change-Id: I951fec5690c1f0d848d86c8859c3e9a171fca887

Former-commit-id: c89bd75a18c53e7205eb0bbfbb4179edcec465c1