Follow-up fixes to MusicRecognitionManagerService
Added a check to only allow the default MusicRecognitionService to be called, unless the caller is the same app that hosts the MusicRecognitionService, as is the case in tests.

Refactored MusicRecognitionManagerPerUserService to not hold the client's callback as a field to prevent leaks and needing to null it out (then requiring null checks).  Instead, it's always a local variable or a final field in ephemeral objects.

Bug: 174744852
Bug: 168696277
Test: atest CtsMusicRecognitionTestCases
Change-Id: I70ff3c38da7096c60f82fe75b47674b4b233e672

Former-commit-id: 5b7c195b0d4c18981f1d2c49bf12200cdc0240d6