Update client pid and priority on requesting resources
The previous implementation queried the client process id only once
during the client registration. But the pid could change before a
client request a resource.

We change the impl to update the pid and priority on request.

Note that the update only happens when the pid changes compared to
the one recorded during client registration or updateClientPriority
API call.

Test: atest com.android.server.tv.tunerresourcemanager
Bug: 176186326
Change-Id: I8d83f7a39d47dc85a00cc4d216080d7eebb66b94

Former-commit-id: 8af4d0bbe9d51a697a6d35fe92a6a401c42f4aa4