Check if mGrContext exists when requiring a vk context.
Vulkan context is shared with HardwareBitmapUploader and the
RenderThread, meaning if the upload thread runs prior to the render
thread then some configuration is skipped such as recording the max
texture size and cache setup, which affects benchmark tools. So we need
to check whether the GrContext exists in addition to checking if the
shared VkManager has been properly intialized.

Bug: 175913056
Test: hwuimacro works for benchmarks that upload HW bitmaps prior to
benchmark execution. For example: tvapp_norcEglImage

Change-Id: I149c294b39f1e88226fc5d8d72c30bc4fb3d8404

Former-commit-id: 7e7c3d72bcd903542b1556efb05dd7512606ba4f