Add log about adding hidden task
When adding task with the same affinity of another existing task
to recents list, the old task may be trimmed. The log will help
to know where the replacing task comes from.

Also fix potential wrong info of Task#toString that some states
are not constant which shouldn't be cached.

Bug: 169507045
Test: Launch multiple activities with the same task affinity
      and flags NEW_TASK | MULTIPLE_TASK.
Change-Id: I712f0c617d31822c7a4828d740d27677ff7e9d52

Former-commit-id: 0812325d3d8d267b9990ba31f8e14b126659a06b