Reduce job readiness checking.
Don't check the full list of jobs whenever we evaluate a single job's
state. Every job that's not ready would be checked anyway.

Note, however, that not checking the list if a job would not be ready
means we will not defer an already set alarm, so we will have more
alarms go off even though we know the job would not be ready at that
time.

Bug: 138469672
Bug: 176987520
Test: atest CtsJobSchedulerTestCases
Test: atest FrameworksMockingServicesTests:JobSchedulerServiceTes
Test: atest FrameworksMockingServicesTests:TimeControllerTest
Change-Id: I41d17ae1df10484f79deb48b197afadbd9cee693

Former-commit-id: 39cc594df4d75dc43ce50e64a78a40eed8cd429d