Unset mNextDrawUseBlastSync if VRI won't draw
There are some checks before attempting to draw in VRI. If we get a
blast sync relayout, but aren't going to draw, unset
mNextDrawUseBlastSync. Otherwise, we end up blocking indefinitely.
There would never be a draw callback so we would never unset that flag.

If there are no pending draws, report back immediately so we don't block
WMS since it could be waiting on the app to draw.

Test: enable shell transitions, unlock device
Bug: 175229445
Change-Id: I8e1f524862e2b9a48949cf6846141d67129c507c

Former-commit-id: d862b98d91bfaeba0c5be6eed23868f20e56884d