Refactor default transit animations into handler for testing
In order to make shell transitions testable, this makes the
default animations into a TransitionHandler. This is also cleaner,
more consistent, and will make expanding the default animations
easier later.

The reason this is needed is because animations themselves aren't
very amenable to testing, so to test the interface logic, we need
to be able to replace the default handler with one that simulates
just the contract of the default handler (don't claim any requests,
but always perform animation).

Also added some tests for both basic transition flow and for
the TransitionHandler contracts

Bug: 169035082
Test: atest ShellTransitionTests
Change-Id: I92e367af155831c11f583322e95151a2cb645caa

Former-commit-id: 1229c01a4e1a3b68f3e7c83f549793e7e6b625d2