Skip sending TaskInfo change while not in pip
When recents animation finishes, it eventually triggers
ActivityRecord.setPictureInPictureParams if the Activity is pip-able.
And in the autoEnterPip case, this happens before the windowing mode
being flipped to PINNED and therefore callbacks the
FullscreenTaskListener and resets the leash to (0, 0).

Since it does not make sense dispatching TaskInfo change
while not in pinned mode, we simply skip it here.

Video: http://rcll/aaaaaabFQoRHlzixHdtY/eFlluOg7dAqjzsS6G8UYoX
Bug: 177073398
Test: see video
Change-Id: I782f94a1b5fe19973f21e1694b8d688575f1b7e8

Former-commit-id: fa10d0a3da937ecaebdba7382662a2f4c1e82622