PinnerService Pin split APKs (behind a flag).
This CL lets PinnerService pin the split APKs, in addition to the
base APK. This matters when the APK is built as AAB.

This behavior is hidden behind "pin_split_apks" DeviceConfig flag,
which is false for now: This is becaue AAB-built apps have
broken pinlist for the base APK and is missing the list for
split APKs, ending up pinning too much. The flag should be flipped
once the these APKs get proper pinlists.

Bug: 174698005
Test: Ran following commands and inspected the output:
  $ adb shell device_config put runtime_native_boot pin_camera true && \
    adb shell device_config put runtime_native_boot pin_split_apks true
  $ adb shell cmd pinner repin && adb shell dumpsys pinner
  $ adb shell device_config put runtime_native_boot pin_camera true && \
    adb shell device_config put runtime_native_boot pin_split_apks false
  $ adb shell cmd pinner repin && adb shell dumpsys pinner
Change-Id: Ic9c82b2538095e53daf9c1ef6557f487e8fac0a4

Former-commit-id: 2a5fcbde13193b02c67fbfe6f6b02e2596026d18