[Attempt #2] Add privileged to BROADCAST_CLOSE_SYSTEM_DIALOGS
Original CL got reverted because target aosp_arm64-userdebug was failing
to boot. Apparently, the right place to put the permission for launcher3
is in frameworks/base/data/etc/com.android.launcher3.xml, not
frameworks/base/data/etc/privapp-permissions-platform.xml (despite it
having an entry on the latter) because launcher3 lives in system_ext.

-- Original CL description

CellBroadcastReceiver, a mainline module signed with the networkstack,
needs it before showing emergency alerts to the user.

This is inline and ok with the rest of of the restrictions given that
STATUS_BAR that allows calls to StatusBarManager.collapsePanels() also
has "privileged" protection level.

Also adding permission to privapp-permissions-platform for Launcher3
because of b/177506944, despite it getting it via "recents" protection
level.

Bug: 177346942
Bug: 159105552
Test: atest PermissionPolicyTest
Test: Run DeviceBootTest for aosp_arm64-userdebug in forrest.
Change-Id: Ie3298c839003aeaca3cc84782e8846374beca5b8

Former-commit-id: 5e390e7998f4aba46c13b1566cdfb8152ee40f0b