Add privileged to BROADCAST_CLOSE_SYSTEM_DIALOGS
CellBroadcastReceiver, a mainline module signed with the networkstack,
needs it before showing emergency alerts to the user.

This is inline and ok with the rest of of the restrictions given that
STATUS_BAR that allows calls to StatusBarManager.collapsePanels() also
has "privileged" protection level.

Also adding permission to privapp-permissions-platform for Launcher3
because of b/177506944, despite it getting it via "recents" protection
level.

Bug: 177346942
Bug: 159105552
Test: atest PermissionPolicyTest
Change-Id: If027c6824a58c8b9a917f9995e34bf8b306cab54

Former-commit-id: 4d62bdba9772a38a13a9e63ea310914adf3acc33