Task letterbox for resizable fixed-orientation app in dual DisplayAreas
Before, when both DisplayAreaGroups set ignoreOrientationRequest to
false, launching two resizable apps with different orientations will
end up with two apps in the same orientations. One app is in the
orientation different from its request.

This is because Task only letterboxes when ignoreOrientationRequest is
true, but it should also letterbox when the TaskDisplayArea can't
specify orientation (happens when it doesn't have focus).

Bug: 176566401
Test: atest WmTests:DualDisplayAreaGroupPolicyTest
Change-Id: I6f093f3c2631970fcc2fa839e5b394106e5acc29

Former-commit-id: 3e4f8e1a7c731a03a587bcf2b478aa40caab2356