Fix fingerprint dialog color contrast for dark theme
The code to set the color of the help text shown on the fingerprint
dialog does not currently check the context theme. This means that the
wrong color is used when the system is using a dark theme.

This commit fixes the issue by ensuring the appropriate themed color is
used.

Test: Manual
1. Turn on system dark theme (Settings > Display > Dark theme)
2. Trigger the fingerprint dialog (e.g. Wi-Fi > Network details > Share)

Bug: 174831630
Change-Id: I41dfb536e72063ae8ea84d51fa0e50613f920090
Merged-In: I41dfb536e72063ae8ea84d51fa0e50613f920090

Former-commit-id: 2a8ab5efe7719336d46761f3fcc925c7fe43b440