Make no-history activity be finished with different tasks
Activities that specify No History must be removed once the user
navigates away from them. Currently, the no-history activity is
managed in Task level which result in the behavior broken when a
fullscreen activity was started on top of the no-history activity
within different tasks.

This CL make the no-history instance cached reference from
TaskSupervisor and support task to manage the behavior. The
no-history activity will be finished only from the same
display with the next top-most opaque activity.

Bug: 168871922
Test: atest ActivityStarterTests

Change-Id: I6d15b0fc6cecbbb7af545c297cd3524c47afe6b4

Former-commit-id: ff2ee7296877469e77b87bb716d78916887c8e14