Remove detachChildren
There's no longer a need to have detachChildren since we only used it to
ensure the children were not deleted before the exit animation. The
timing when closing an app has changed so the app is not stopped until
everything has closed. This means the children of the window will not
try to remove themselves until after the animation completes

Test: Close app with SurfaceView. No hole in exit animation
Bug: 177557720

Change-Id: I48642864d5315b2bf563c675756283fa73554de4

Former-commit-id: ce7428972d302908dfc08d1712e112ebe849796b