Add traces for activity create/resume/pause/stop/destroy
With component name, it is clearer to know which activities
are executed. And the traces are closer to the implementation
of app, so the relative time between app and framework can be
shown by overlapping with the traces of transaction items.

Bug: 175028240
Test: build
Change-Id: I76faf6265f38c211ffd69a36b6e3ad882465251e
(cherry picked from commit d7af5e3888a7916e9911e9ef35cbfcc3bdce8520)

Former-commit-id: f7c9bd991a1da1160dabb82616a44813a61aa008