Ensure DeviceStateProviderImpl considers first state before others.
DeviceStateProviderImpl#notifyDeviceStateChangedIfNeeded() would skip
the first state and consider other states, before falling back to the
first state if no others match. This fails to handle the case where a
state is declared in the config file without any conditions to allow the
state to be requested externally (for ex, TENT mode).

Fixes: 178387437
Test: atest DeviceStateProviderImplTest
Change-Id: I7240c412b98978028c3d86257723cf00dba28ea0

Former-commit-id: 12b4ecdb9d0ed3d604f7522457653b20d149df7b