Remove preserve surfaces
When using blast, there's no need to call preserve surfaces since the
client can synchronize layer changes with the incoming buffer in the
new size. If the client uses the WindowOrganizer.applySync call they can
ensure they only update the layer geometry in the same transaction that
the new buffer is submitted in.

When removing preserveSurfaces, we can also remove reparentChildren
since that was the only caller of reparentChildren

Test: Split with SurfaceView
Bug: 177557720
Change-Id: I4574ac0d3b8a63b13ac44846e729b73ca0f88f23

Former-commit-id: 32f2817bbdca2e8b1953aec7ef6784ae94d50775