Fix SensorRect location
The window includes the additional padding, but SensorRect was
still positioned flush to the top+left. This CL 1) Translates the
SensorRect and positions it correctly within the window, and
2) Removes translation from enrollment animation, since it's already
taken care of by SensorRect's fixed positioning

Without this change, the enrollment animations are in the right
place, but the HBM dot was wrong (only slightly, due to the small
padding so far). However, with larger padding, the dot would be
really off.

Bug: 179083247
Test: manual
Change-Id: I556f5bebe5177a748f19cadeab86d1d962d3e1f1

Former-commit-id: 8e24475f1eae4974785a9aced9a7048c2e841de6