Remove an unused runtime flag "debug.optimize_startinput"
This is a logical revert of a CL [1] that aimed to explore an approach on
optimizing onStartInput(), which has never been enabled by default.

Since the flag has never been enabled, there should be no user
observable behavior change in this CL.

This CL also removes InputMethodManagerService.DebugFlag that was
introduced in the same CL with an assumption that we may most likely
want to consider using platform-wide infrastructures such as
android.app.compat.CompatChanges before trying to reinvent our own
infrastructure in 2021.

 [1]: Ibf9dab3d9c138b5f04e053d41ee4fd248c78e4da
      3ffed59167d86eb8f959c390722a71b134ce02ac [formerly 75a92392218a12eaabd2187724f7175ef38d362b]

Bug: 37617707
Bug: 171827834
Test: presubmit
Change-Id: Ie7b09f16e517c0550ac495a8a8c8ef20abb6938e

Former-commit-id: 8123d67ce8ce4e06b6a1e26b6182b0edf1093df6