Only wait for significant windows to unfreeze display
This may reduce frozen time by ~60% (~100ms on a mid-end device)
when rotating device with a simple activity that can handle
orientation change (no relaunch).

The unfreezing procedure still waits for the windows of Activity,
IME, wallpaper. Other windows such as navigation bar, status bar,
toast, overlay, they are hidden when rotation animation is set for
freezing screen. And if they are redrawn in new rotation or reach
timeout, their windows will fade in. If the redraw time is quick
enough, the visual appearance will be almost the same as waiting
for all windows because the the duration of rotation animation make
the fade-in animation unnoticeable.

Bug: 178472794
Test: atest DisplayContentTests#testHybridRotationAnimation
Test: Rotate device and check log "Screen frozen for".

Change-Id: I57cce881a2d685a2f1130d7e9a335fd9d86ba360

Former-commit-id: 1a7dc4c24cab0e80ba302b2b8693b1ad16166b73