Apply launch power mode during freezing screen
The most common case of freezing screen is display orientation
change. It usually needs to relaunch activity and redraw windows,
that is similar to the scenario of normal activity launch. So
also apply the "launch" power mode to speed up the procedure.

This can reduce about 10% freezing time for both relaunch and
no-relaunch cases.

Bug: 178472794
Test: Rotate device and check log "Screen frozen for".

Change-Id: Idd4a2374d8ce978af36d75135b680a10c240f66d

Former-commit-id: 4bc0e4eb1250d6eccfb8d4404f7bc22001eb61fd