Fix ellipsis reinitialization for static cached StaticLayout.
DynamicLayout uses a static StaticLayout instance internally to avoid
allocations. For the specific case of multiline, TruncateAt.START, and
no ellipsization needed (may also be triggered by slightly different
params), the ellipsis results for a reused static instance are not
modified. If the instance was previously ellipsized, the new result
keeps that ellipsization result even though it shouldn't be ellipsized
(at least for some lines).

Bug: 179524847
Test: atest DynamicLayoutBlocksTest DynamicLayoutTest TextLayoutTest \
  TextViewTest
Change-Id: I1b4142648ec1b6b3fc2ece0ec55aac9539497134

Former-commit-id: 5b873f1366e84241015a81786761ef147790bf9b