Remove retry in Fingerprint21 getService
The retry was preventing the system from booting on devices that
were failing to load the fingerprint HAL.

Instead of retrying immediately, the system will try to retrieve the HAL
again only when it's needed. On a misconfigured or bad device where the
HAL cannot be retrieved, the system will simply log an error and move
on.

The retry was originally introduced to fix the unreliable isUdfps check.
Sometimes the HAL wouldn't be loaded early enough, and that would cause
the system to think that the sensor is not a UDFPS, as a side effect.

The check is fixed in this change by relying on the presence of a UDFPS
config in config.xml.

Bug: 178053992
Bug: 179465449
Test: the absence/presence of UDFPS is detected correctly
Test: the system boots with manually deleted HAL binary
Change-Id: I5d3b81e26f6ed421d32691c6cab55ef90f4b5ec8

Former-commit-id: 8d2fd57600e2c19b5e32bb5a21a625a5fe781843